Fix `pl.debug_print` passing extraneous scalar placeholders to TPU logger.
When the Python-side format string contains text after the last scalar placeholder, e.g. `x={} y={} zzz` instead of just `x={} y={}`, an extra `$2` placeholder was appended to the end of the C++-side format string.

As a fly-by, this patch also prepares for (but does not enable) supporting the following format specifiers:
- `{:s}` for signed i32 as decimal
- `{:u}` for unsigned i32 as decimal
- `{:x}` for unsigned i32 as hexadecimal

Float32 scalars seems more involving than simply passing the specifier through the format string and is therefore omitted from this patch.

Reason for not enabling the TPU-specific format specifiers: The format validation codepath is currently not specialized for TPU; we don't want to blanket-enable here.

PiperOrigin-RevId: 822326844